Optimize the integer type usages

+++ dojo/src/models/food.cairo
@@ -9,7 +9,7 @@ pub struct Food {
     #[key]
     pub id: u8,
     pub name: felt252,
-    pub amount: u32,
+    pub amount: u8,
 }
 
 
@@ -27,13 +27,13 @@ mod tests {
             player: player_address,
             id: 1_u8,
             name: 'Apple',
-            amount: 5_u32,
+            amount: 5_u8,
         };
 
         assert_eq!(food.player, player_address, "Player address should match");
         assert_eq!(food.id, 1_u8, "Food ID should be 1");
         assert_eq!(food.name, 'Apple', "Food name should be 'Apple'");
-        assert_eq!(food.amount, 5_u32, "Food amount should be 5");
+        assert_eq!(food.amount, 5_u8, "Food amount should be 5");
     }
 
     #[test]
@@ -45,14 +45,14 @@ mod tests {
             player: player_address,
             id: 1_u8,
             name: 'Apple',
-            amount: 5_u32,
+            amount: 5_u8,
         };
 
         let banana = Food {
             player: player_address,
             id: 2_u8,
             name: 'Banana',
-            amount: 3_u32,
+            amount: 3_u8,
         };
 
         assert_eq!(apple.player, banana.player, "Both foods should belong to same player");
@@ -70,14 +70,14 @@ mod tests {
             player: player1_address,
             id: 1_u8,
             name: 'Apple',
-            amount: 5_u32,
+            amount: 5_u8,
         };
 
         let player2_food = Food {
             player: player2_address,
             id: 1_u8,
             name: 'Apple',
-            amount: 3_u32,
+            amount: 3_u8,
         };
 
         assert!(player1_food.player != player2_food.player, "Foods should belong to different players");
@@ -94,9 +94,9 @@ mod tests {
             player: player_address,
             id: 1_u8,
             name: 'Apple',
-            amount: 0_u32,
+            amount: 0_u8,
         };
 
-        assert_eq!(food.amount, 0_u32, "Food amount should be 0");
+        assert_eq!(food.amount, 0_u8, "Food amount should be 0");
     }
 }
