+++ dojo/src/tests/test_food.cairo
@@ -76,7 +76,7 @@ mod tests {
         println!("Initial Status - Energy: {}, Hunger: {}", 
         initial_status.energy, initial_status.hunger);
 
-        let mut counter: u32 = 0;
+        let mut counter: u8 = 0;
         while counter < 10 {
             // Decrease stats
             actions_system.decrease_status();
@@ -108,7 +108,7 @@ mod tests {
         actions_system.set_current_beast(1);
 
         // Kill beast by decreasing stats
-        let mut counter: u32 = 0;
+        let mut counter: u8 = 0;
         loop {
             let status: BeastStatus = world.read_model(1);
             if !status.is_alive {
@@ -147,7 +147,7 @@ mod tests {
         actions_system.set_current_beast(1);
 
         // We decrease the stats to verify that they increase after feeding
-        let mut counter: u32 = 0;
+        let mut counter: u8 = 0;
         while counter < 20 {
             // Decrease stats
             actions_system.decrease_status();
